Convert DlpContentRestriction and SystemFeature enum to enum class.
Otherwise, if both are included than there is kMaxValue collision.
Also, enum classes are preferable over just enums.

Bug: 1311441
Test: Trybots.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3560459
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987036}

+++ chrome/browser/chrome_content_browser_client_unittest.cc
@@ -520,7 +520,7 @@ TEST_F(ChromeContentSettingsRedirectTest, RedirectSettingsURL) {
   EXPECT_EQ(settings_url, dest_url);
 
   base::Value list(base::Value::Type::LIST);
-  list.Append(policy::SystemFeature::kBrowserSettings);
+  list.Append(static_cast<int>(policy::SystemFeature::kBrowserSettings));
   testing_local_state_.Get()->Set(
       policy::policy_prefs::kSystemFeaturesDisableList, std::move(list));
 
@@ -538,7 +538,7 @@ TEST_F(ChromeContentSettingsRedirectTest, RedirectOSSettingsURL) {
   EXPECT_EQ(os_settings_url, dest_url);
 
   base::Value list(base::Value::Type::LIST);
-  list.Append(policy::SystemFeature::kOsSettings);
+  list.Append(static_cast<int>(policy::SystemFeature::kOsSettings));
   testing_local_state_.Get()->Set(
       policy::policy_prefs::kSystemFeaturesDisableList, std::move(list));
 
@@ -561,7 +561,7 @@ TEST_F(ChromeContentSettingsRedirectTest, RedirectScanningAppURL) {
   EXPECT_EQ(scanning_app_url, dest_url);
 
   base::Value list(base::Value::Type::LIST);
-  list.Append(policy::SystemFeature::kScanning);
+  list.Append(static_cast<int>(policy::SystemFeature::kScanning));
   testing_local_state_.Get()->Set(
       policy::policy_prefs::kSystemFeaturesDisableList, std::move(list));
 
@@ -578,7 +578,7 @@ TEST_F(ChromeContentSettingsRedirectTest, RedirectCameraAppURL) {
   EXPECT_EQ(camera_app_url, dest_url);
 
   base::Value list(base::Value::Type::LIST);
-  list.Append(policy::SystemFeature::kCamera);
+  list.Append(static_cast<int>(policy::SystemFeature::kCamera));
   testing_local_state_.Get()->Set(
       policy::policy_prefs::kSystemFeaturesDisableList, std::move(list));
 
@@ -595,7 +595,7 @@ TEST_F(ChromeContentSettingsRedirectTest, RedirectHelpURL) {
   EXPECT_EQ(GURL("chrome://settings/help"), dest_url);
 
   base::Value list(base::Value::Type::LIST);
-  list.Append(policy::SystemFeature::kBrowserSettings);
+  list.Append(static_cast<int>(policy::SystemFeature::kBrowserSettings));
   testing_local_state_.Get()->Set(
       policy::policy_prefs::kSystemFeaturesDisableList, std::move(list));
